As usual, studying CS related things turned out to be a waste of time. It feel like I mostly learn buzzwords from those things and all the practically useful thing I've learned have been things I've discovered myself.
Principles in Program Analysis starts out damn heavy. No way I am getting through it.

Today one thing that I did was open two issue on the Futhark repo to track the status of the Cuda backend and the FFI.

I've made my determination to do my part in its story, but presently the reality is that I do not believe in the language.

But I do believe in effort and desire. I am not sure how much faith in Athas I should put in, but it is impossible to move forward merely by looking at the ground or the sky, you need to face forward, so I will believe that he will make the right design choices so that Futhark becomes useful in the real world and not just a source of academic papers.

I tend to say hurtful things sometimes, like I do not want to see lists ever again last time, and now I cast doubt on Futhark's high level orientantion and the way Athas signals that he wants Futhark to be everything, but those are my true feelings.

Every once in a while I would think about the language because I have decided to make it a obligation, but from now until the time Athas surprises me by completing the Cuda backend and the FFI, I will not give it another thought as a matter of principle. I will assume those things will not be made and not stress about the two issues.

And when it is made, it won't be hard to do and it will be great.

Today when I started writting the thing it was 10am and after breakfast when I finish the thing it was 2pm so that long rambling essay I posted on the Cuda issue took quite a chunk out of my day.