Today most of my time seems to have been taken up by trying to come to a decision regarding Futhark.
Just now I've decided.

I do not want Futhark six months from now. The time to strike is now. The time when it would be useful is now. The time when I would be able to make use of it is as soon as I finish the allocator, though I would need to deal with the backend, but it would be worth it.

I do not want Futhark six months from now when the Spiral's ML library will be complete. I want it now when the time to plug it in is ideal and the energy is at its peak.

For the past year, I've sort of imagined Spiral would be a baseline for Futhark to be compared against, but I think now in the moment of truth is that Spiral is pretty overwhelming.

I am going to push Athas to give me an ETA and decide whether to keep or release my obligation towards Futhark based on that.

I intend to come to a decision in the next two days at most.

I honestly, had not intended to make this post today at all, or do my public interactions in such bursts, but riding the surge of inspiration is how I do things.

If Futhark is not ready, then it is simply not ready - it is not my responsibility. Hindisght it 20/20.

Spiral is aaaalmost ready.

I would say that I am willing to wait for about two weeks from the point Spiral is trully ready for Athas to make the first version of the codegen.

So without a doubt, I would like to start work on the Futhark backend in the next 3-4 weeks.

...Well, that is about it for the day for me. It is 7:40pm for God's sake.

Now that I feel this is resolved and my studies are done as well, tomorrow I will sharpen my focus. I mostly have the new allocation scheme planned out and it won't be any trouble to implement it.

One way or another, I will make the ML library.

My prediction though for how this will go is not optimistic - generally positive interaction take time to be crested and I am kind of jumping right in here. Given that there even wasn't the Cuda backend on the issues page, that is a good indication that the thing was not a priority for him so it is highly unlikely that he will say something like less than a month.