I am up.
I just now fixed the module_map so it does not repack it. Though maybe I should just add a special unbox instead?

Actually let me go with that...

Yesterday Athas said something interesting that Futhark is really good at optimizing unused arguments.

Even in the face of recursion? I wanted to ask him that, but I pretty much figured out how to do the optimization completely now, in the last few hours.

This is something that gave me significant trouble before.

I am not sure whether I am lucky or not.

I will need this optimization to deal with unused arguments when doing maps where I need to pass in both primals and adjoints, but forget that for now.

Yesterday I said that studying CS was a waste of time as usual, but I found something interesting in the bunch. The first was the tagless GC paper.

http://wiki.c2.com/?TaglessGarbageCollection

The other would be the project Lancet paper by Tiark Rompf.

https://github.com/TiarkRompf/lancet
https://www.cs.purdue.edu/homes/rompf/papers/rompf-pldi14.pdf

Now, I do prefer my own way of doing staging more than the way LMS does this.

Something like this would be of great benefit to Spiral, too bad this is JVM only. I might look around whether there is something similar to this on the .NET.

Well, I just now added LayoutToNone. Let me read the paper first and then I will move on to the allocator.