With this I am finally back to parity. The new macros work and I put in the binary operator one.
Fuuuu...I feel a bit low energy after being dumped by Futhark. I know, I know, it was just too soon, the timing was off. I got attached to it due to its elegance, but I cannot marry it with Spiral. Not now and probably never. The next time we see each other we will have to fight it out.

I admit, based on how things were going, I really misunderstood Athas' intentions and thought he had plans to turn it into a general purpose language eventually.

Academics really do have different incentives than bandits such as myself.

That whole interaction could have been:
"Pls give me an ETA for the backend?"
"6 months plus."
"Thx, bye."

I never saw this in other people, but I sometimes have the tendency to develop an affection for concepts and abstractions, and in this case a language.

I will admit though, while I like what Futhark represents, I do not trust that it can achieve its ideals with its current design.

If by some chance it had the Cuda backend, I would have had to go through some pains to get it work well. I would have done it no doubt, but there are quite a bit of things I did not like in the generated code such as that it had trouble propagating sizes (or maybe types?) and that it plopped a huge pile of unused code directly into the source.

Futhark could definitely benefit from having Spiral inside it...how lewd. I meant to say a Spiral-like language internally.

Ultimately, what the guys working on staging are doing is the future of compilation. I'll put my faith in that all the way. Compared against low level languages, Spiral is an improvement on them in every way imaginable.

And probably dynamic staging of the sort it is doing is the most ergonomic way of doing staging in a language and the optimal design, though no doubt I will find plenty of objections to this.

What I could really use is something like LLVM except for high level optimizations. I had wanted Futhark to fill that role.

Maybe I could consider compiling to TF or something like that? Or maybe to PyTorch? Though I dislike those things and would prefer to kill them instead.

Theano is dead, but I think I will give it a shot at competing with Google and Facebook head on.

The fact that I have to use macros here is annoying.

If I had something like Lancet on the .NET and better .NET interop, I could take the language to a completely new level.

For example, there is no real reason why literals have to be only primitive types. They could potentially be any object. Having something like that in the language could be quite powerful, though I would need to take care to restrict the datatypes so they are immutable.

...That would be a problem. Cancel that, literals cannot be any object, just immutable ones.

So based on that, rather than Racket being the ideal platform, the ideal compiler for a staged language would be Haskell?

No, I still need mutable dictionaries inside the language to implement it, and Haskell cannot support dynamic typing.

So the ideal platform would be a dynamic language with a strict separation between immutable and mutable objects?

Actually, mutable objects could be supported, but they need to be Copyable. .NET and JVM's opaque objects are no good for this.

This is the essential condition for staging - copyability.

...Enough of this.

I do not feel like starting on the allocator, but let me sketch it out just for starters.