Let me save here.
What I should do next is using the allocator make a Tensor module linked to it and then do the map example with it.

When that is done I will remove the thing currently in the library and replace it with the allocator using one.

I think for now I will call it a day. It is good that I did this much at least. After all the studies of the last few days I nee to get back into grove of writting code. Today I spent about 3 hours just studying the paper by Rompf.

Before I really stop, I will look a little if there is something like Lancet for the .NET, but I am not expecting anything.