Jesus, I am still at it.
But I did it. I am finally done and the thing works. I even optimized the shit out of the allocate function.

There is some pretty brilliant CPS work in there. It took me a bit to evolve it to that point.

I was on fire today. What I am going to tomorrow is CPS all the CudaTensor allocating functions and them do lexically scoped management of resources.

This is what I should have done last time with the library, but I guess my skills were not up to par to think of such an idea back then.

It is quite a simple and elegant idea though, and the particular method of allocating nodes should be quite sufficient for most neural net uses. In the present time I would go as far as to say all of them in fact.

It definitely has much more quality than the thing I used last time which was to keep a list of nodes and resize them upwards. Instead of such a scheme right now have a stack which should ensure that no fragmentation occurs as the batch sizes change from run to run.

Since it is getting late, let me stop here for the day. Tomorrow I will compile the compiler into a dll, finished the allocator maybe put in gemm.

What I will be able to do after that is start work on doing basic automatic differentiation. It has been over a year since I did ML last.

This time it will be different. My skills have leveled up and I have the best possible language in existence to do it with.

If I cannot draw out the true power of the machines, then nobody will be able to. I will not hesitate to go through the passage that preys upon the weak.